+++ plugins/woocommerce-admin/includes/api/class-wc-admin-rest-leaderboards-controller.php
@@ -92,7 +92,7 @@ class WC_Admin_REST_Leaderboards_Controller extends WC_REST_Data_Controller {
 				),
 				$persisted_query
 			);
-			$coupon_url  = wc_admin_url( 'analytics/coupons', $url_query );
+			$coupon_url  = wc_admin_url( '/analytics/coupons', $url_query );
 			$coupon_code = isset( $coupon['extended_info'] ) && isset( $coupon['extended_info']['code'] ) ? $coupon['extended_info']['code'] : '';
 			$rows[]      = array(
 				array(
@@ -158,7 +158,7 @@ class WC_Admin_REST_Leaderboards_Controller extends WC_REST_Data_Controller {
 				),
 				$persisted_query
 			);
-			$category_url  = wc_admin_url( 'analytics/categories', $url_query );
+			$category_url  = wc_admin_url( '/analytics/categories', $url_query );
 			$category_name = isset( $category['extended_info'] ) && isset( $category['extended_info']['name'] ) ? $category['extended_info']['name'] : '';
 			$rows[]        = array(
 				array(
@@ -223,7 +223,7 @@ class WC_Admin_REST_Leaderboards_Controller extends WC_REST_Data_Controller {
 				),
 				$persisted_query
 			);
-			$customer_url = wc_admin_url( 'analytics/customers', $url_query );
+			$customer_url = wc_admin_url( '/analytics/customers', $url_query );
 			$rows[]       = array(
 				array(
 					'display' => "<a href='{$customer_url}'>{$customer['name']}</a>",
@@ -288,7 +288,7 @@ class WC_Admin_REST_Leaderboards_Controller extends WC_REST_Data_Controller {
 				),
 				$persisted_query
 			);
-			$product_url  = wc_admin_url( 'analytics/products', $url_query );
+			$product_url  = wc_admin_url( '/analytics/products', $url_query );
 			$product_name = isset( $product['extended_info'] ) && isset( $product['extended_info']['name'] ) ? $product['extended_info']['name'] : '';
 			$rows[]       = array(
 				array(