+++ plugins/woocommerce-admin/includes/core-functions.php
@@ -37,10 +37,10 @@ function wc_admin_number_format( $number ) {
 function wc_admin_url( $path, $query = array() ) {
 	if ( ! empty( $query ) ) {
 		$query_string = http_build_query( $query );
-		$path         = $path . '?' . $query_string;
+		$path         = $path . '&' . $query_string;
 	}
 
-	return admin_url( 'admin.php?page=wc-admin#' . $path, dirname( __FILE__ ) );
+	return admin_url( 'admin.php?page=wc-admin&path=' . $path, dirname( __FILE__ ) );
 }
 
 /**