+++ plugins/woocommerce-admin/packages/components/src/link/index.js
@@ -3,40 +3,47 @@
  * External dependencies
  */
 import { Component } from '@wordpress/element';
-import { Link as RouterLink } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 /**
  * WooCommerce dependencies
  */
-import { getAdminLink } from '@woocommerce/navigation';
+import { getAdminLink, getHistory } from '@woocommerce/navigation';
 
 /**
  * Use `Link` to create a link to another resource. It accepts a type to automatically
  * create wp-admin links, wc-admin links, and external links.
  */
 class Link extends Component {
+	// @todo Investigate further if we can use <Link /> directly.
+	// With React Router 5+, <RouterLink /> cannot be used outside of the main <Router /> elements,
+	// which seems to include components imported from @woocommerce/components. For now, we can use the history object directly.
+	wcAdminLinkHandler( e ) {
+		e.preventDefault();
+		getHistory().push( e.target.closest( 'a' ).getAttribute( 'href' ) );
+	}
+
 	render() {
 		const { children, href, type, ...props } = this.props;
-		if ( this.context.router && 'wc-admin' === type ) {
-			return (
-				<RouterLink to={ href } { ...props }>
-					{ children }
-				</RouterLink>
-			);
-		}
 
 		let path;
 		if ( 'wp-admin' === type ) {
 			path = getAdminLink( href );
-		} else if ( 'external' === type ) {
-			path = href;
 		} else {
-			path = getAdminLink( 'admin.php?page=wc-admin#' + href );
+			path = href;
+		}
+
+		const passProps = {
+			...props,
+			'data-link-type': type,
+		};
+
+		if ( 'wc-admin' === type ) {
+			passProps.onClick = this.wcAdminLinkHandler;
 		}
 
 		return (
-			<a href={ path } { ...props }>
+			<a href={ path } { ...passProps }>
 				{ children }
 			</a>
 		);