+++ plugins/woocommerce-admin/packages/navigation/CHANGELOG.md
@@ -1,4 +1,10 @@
+# 3.0.0 (unreleased)
+
+- `getHistory` updated to reflect path parameters in url query.
+- `getNewPath` also updated to reflect path parameters in url query.
+
 # 2.1.0
+
 - New method `getSearchWords` that extracts search words given a query object.
 - Bump dependency versions.